fix: Price의 잘못된 생성자 타입 변경

+++ src/main/java/kitchenpos/domain/Price.java
@@ -20,7 +20,7 @@ public class Price {
         value = null;
     }
 
-    public Price(int value) {
+    public Price(long value) {
         this(BigDecimal.valueOf(value));
     }
